backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
astor
pytest
